better version of exhausting a reader into a channel
