+++ pkg/dhpsi/dhpsi_ristretto.go
@@ -13,6 +13,7 @@ import (
 const (
 	RistrettoTypeGR = iota
 	RistrettoTypeR255
+	RistrettoNative
 )
 
 type Ristretto interface {
@@ -26,6 +27,10 @@ type GR struct {
 
 type R255 struct {
 	key *r255.Scalar
+}
+
+type Native struct {
+	key [EncodedLen]byte
 }
 
 func NewRistretto(t int) (Ristretto, error) {
@@ -89,3 +94,5 @@ func (r *R255) Multiply(dst *[EncodedLen]byte, src [EncodedLen]byte) {
 	tmp = p.Encode(tmp)
 	copy(dst[:], tmp)
 }
+
+// built-in
